Actualización codigo para pruebas: CU06 y CU07

+++ 04. Implementación/02. Pruebas/tempus_web/modelos/mesas/Docente.php
@@ -1,5 +1,4 @@
 <?php
-
 /**
  * Esta clase corresponde con la tabla Docente de la base de datos.
  * Contiene los metodos necesarios para crear, borrar, buscar o modificar un docente.
@@ -74,7 +73,20 @@ class Docente
      * */
     public function setNombre($nombre)
     {
-        $this->nombre=$nombre;
+        $nombre = Utilidades::convertirCamelCase($nombre);
+        $this->nombre = $nombre;
+    }
+    
+    /**
+     * Asigna los datos indicados por parametro a los atributos de clase.
+     * @param integer $iddocente Identificador del docente.
+     * @param string $nombre Nombre del docente.
+     * */
+    public function cargar($iddocente, $nombre)
+    {
+        $nombre = Utilidades::convertirCamelCase($nombre);
+        $this->iddocente = $iddocente;
+        $this->nombre = $nombre;
     }
     
     /**
@@ -85,12 +97,15 @@ class Docente
      * */
     public function crear($nombre)
     {
-        $this->buscar($nombre);
-        if (is_null($this->iddocente)) {
-            ObjetoDatos::getInstancia()->ejecutarQuery("INSERT INTO docente VALUES (null,'".$nombre."')");
-            if (ObjetoDatos::getInstancia()->affected_rows > 0) {
-                $this->iddocente = (Int) ObjetoDatos::getInstancia()->insert_id;
-                $this->nombre = $nombre;
+        if ($nombre) {
+            $this->buscar($nombre);
+            if (is_null($this->iddocente)) {
+                $nombre = Utilidades::convertirCamelCase($nombre);
+                ObjetoDatos::getInstancia()->ejecutarQuery("INSERT INTO docente VALUES (null,'{$nombre}')");
+                if (ObjetoDatos::getInstancia()->affected_rows > 0) {
+                    $iddocente = (Int) ObjetoDatos::getInstancia()->insert_id;
+                    $this->cargar($iddocente, $nombre);
+                }
             }
         }
     }
@@ -102,35 +117,31 @@ class Docente
      * */
     public function borrar($iddocente)
     {
-        $consulta = "DELETE FROM docente WHERE iddocente = ".$iddocente;
-        ObjetoDatos::getInstancia()->ejecutarQuery($consulta);
-        if (ObjetoDatos::getInstancia()->affected_rows > 0) {
-            $this->iddocente = null;
-            $this->nombre = null;
+        if($iddocente) {
+            $consulta = "DELETE FROM docente WHERE iddocente = ".$iddocente;
+            ObjetoDatos::getInstancia()->ejecutarQuery($consulta);
+            if (ObjetoDatos::getInstancia()->affected_rows > 0) {
+                $this->cargar(null, null);
+            }
         }
     }
     
     /**
-     * Busca a un docente en la base de datos utilizando su nombre. Si el docente
-     * se encuentra cargado en la base de datos se actualizan los atributos id y 
-     * nombre. En caso contrario, los atributos serán nulos.
-     * La búsqueda en la base de datos se realiza igualando el parametro ingresado
-     * con el campo nombre de la tabla docente. No se obtienen similares.
+     * Busca a un docente en la base de datos utilizando su nombre. Si el docente se encuentra 
+     * cargado en la base de datos se actualizan los atributos id y nombre. En caso contrario, los
+     * atributos serán nulos. La búsqueda en la base de datos se realiza igualando el parametro 
+     * ingresado con el campo nombre de la tabla docente. No se obtienen similares.
      * @param $nombre String Recibe el nombre del docente a buscar.
      * */
     public function buscar($nombre)
     {
-        $consulta = "SELECT * FROM docente WHERE nombre = '".$nombre."' LIMIT 1";
+        $consulta = "SELECT * FROM docente WHERE nombre='{$nombre}' LIMIT 1";
         $this->datos = ObjetoDatos::getInstancia()->ejecutarQuery($consulta);
         if ($this->datos->num_rows > 0) {
-            /* Se ha encontrado un docente que cumple la condicion de busqueda */
             $fila = $this->datos->fetch_row();
-            $this->iddocente = $fila[0];
-            $this->nombre = $fila[1];
+            $this->cargar($fila[0], $fila[1]);
         } else {
-            /* No se ha encontrado un docente */
-            $this->iddocente = null;
-            $this->nombre = null;
+            $this->cargar(null, null);
         }
         $this->datos = null;
     }
